requirements.txt
streamlit==1.54.0
gspread==6.1.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.160.0
python-dotenv==1.0.1
pillow==11.1.0
cloudinary==1.44.1
pandas==2.2.2
